requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
openai>=1.50.0
